ovs_performance: don't require virt info for p2p tests
Currently, a number of the configuration options are required, even if
skipping all the tests that need them.  This is true especially of the
vm-related information, when running with --skip-pv-test --skip-pvp-test
and --run-pp-test.

This commit loosens the checks to allow these flags to be specified to
run the P2P testing without any VM.

Signed-off-by: Aaron Conole <aconole@redhat.com>