ovs_perf: Conversion Python2 to Python3
There is need for a Python2 and Python3 compatible ovs_perf scripts,
since many distributions have been switching to default with Python3.

The main difference between Python2 and Python3 is the basic types
that exist to deal with texts and bytes. Python3 introduces more
distinguishing string types, but that means the "str" usage for
"bytes-on-the-wire" type programming will run into errors doing
comparisons.

Some of this conversion was provided by the "2to3" utility.

All the requested changes from the pull requests have been
incorporated in this commit.

Note: The translate function isn't Python3 backwards compatible and
the fix is tedious and looks ugly. So I have used the replace function
instead which is Python3 backwards compatible.

Signed-off-by: Nandan Kulkarni <nakulkar@redhat.com>