ovs_performance: Adds support for the OVS NORMAL rule
This change allows testing with the Open vSwitch NORMAL rule, aka FDB
learning. The difference with the normal L2 PVP test is that it will
only add a single OVS OpenFlow rule (FDB learning).

To use the NORMAL mode add the following additional parameters to your
PVP setup:

  --flow-type=L2-NORMAL --mac-swap

Currently this is only supported for the PVP scenario with L2
addresses.