ovs_performance: Add support for DPDK 18.11
With DPDK 18.11 we need to use the legacy-mem mode, or else
zero packet loss is not reached. In addition normal pvp results are
also lower.

In addition we need to wait for testpmd to shutdown, or else the next
test might fail due to huge memory not yet being released.