scripts: Remove all scripts from this repository
This commit removes all scripts from this repository. Reason being
that this is only a documentation change to show what is needed to
make this work on RHEL8. At a later stage we will try to integrate
these documentation changes into the main repository.

To avoid syncing the scripts with the master branch every time there
is a change they are removed.