ovs_performance: Add simple check to wait till testpmd is ready
Rather than a fixed 2 second delay, a simple bash script is added to
wait for testpmd to be running on the VM. This might not be the best
solution but it works without adding additional test scripting on the
VM itself.

In addition, an option --testpmd-startup-delay has been added to use a
fixed time delay. Note that if you make this delay to long (about 2
seconds longer than the actual delay needed, the VMs tx/rx ppm rate
might be wrongly calculated.